LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-noscroll-body-xhidden-expected.txt

This tests that a page cannot be scrolled horizontally with touch (but can still be scrolled vertically) if its body has style overflow-x:hidden. The scroll events in this test have both an x and y component.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
PASS document.body.scrollTop is 110
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS document.body.scrollTop is 205
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
